+++ droit/loader.py
@@ -7,8 +7,35 @@
 
 from . import models as _models
 from . import legacy as _legacy
+from . import analyzer as _analyzer
 
 from typing import List as _List
+
+
+def nonEscapeSplit(string: str, divider: str) -> list:
+	"""Split a string if the divider isn't escaped."""
+	if(len(divider) == 1):
+		resp = [""]
+		if(len(string) > 1):
+			if(string[0] == divider):
+				resp.append("")
+			else:
+				resp = [string[0]]
+			for i in range(1, len(string)):
+				if(string[i] == divider and string[i-1] != "\\"):
+					resp.append("")
+				else:
+					resp[-1] += string[i]
+		else:
+			resp = [string]
+		return resp
+	else:
+		print("error: python-droit: divider has to be a single character")
+		return False
+
+
+def escapeCharacters(string: str) -> str:
+	return string.replace("\\\\", "\\").replace("\\!", "!").replace("\\:", ":").replace("\\-", "-").replace("\\>", ">")
 
 
 def parseLegacy(filename: str, plugins) -> _List[_models.DroitRule]:
@@ -54,79 +81,78 @@ def parseLegacy(filename: str, plugins) -> _List[_models.DroitRule]:
 	return rules
 		
 
-def parseScript(filename: str, plugins=False) -> _List[_models.DroitRule]:
+def parseScript(filename: str, plugins=False, legacyValid=False, warnings=True) -> _List[_models.DroitRule]:
 	plain = open(filename, "r").read()
-	return parseScriptString(plain, plugins=plugins)
+	return parseScriptString(plain, plugins=plugins, legacyValid=legacyValid, warnings=warnings)
 
-def parseScriptString(string: str, plugins=False) -> _List[_models.DroitRule]:
+def parseScriptString(string: str, plugins=False, legacyValid=False, warnings=True) -> _List[_models.DroitRule]:
 	plain = string.split("\n")
 	rules = []
 	
 	for line in plain:
-		if(_legacy.isValidLine(line)):
+		if(legacyValid):
+			isValid = _legacy.isValidLine(line)
+		else:
+			isValid, info = _analyzer.isValidLine(line, infos=True)
+		
+		if(isValid):
 			rule = _models.DroitRule([], [])
 			inp, out = line.split("->")
 
-			inp = inp.split(":")
-			out = out.split(":")
+			inp = nonEscapeSplit(inp, ":")
+			out = nonEscapeSplit(out, ":")
 
 			for inpx in inp:
-				inpx = inpx.split("!")
+				inpx = nonEscapeSplit(inpx, "!")
 				inpx[0] = inpx[0].replace("NOTX", "TEXT*true")
 				
 				if(inpx[1] != ""):
-					children = inpx[1].replace("&arz;", "!").replace("&dpp;", ":").split(",")
+					children = escapeCharacters(inpx[1]).split(",")
 				else:
 					children = []
 
 				if("*" in inpx[0]):
 					attr = {}
-					block, atnm = inpx[0].split("*")
-					block = block.upper()
+					parts = inpx[0].split("*")
+					block = parts[0].upper()
 
 					if(plugins):
 						for plugin in plugins:
 							if(plugin.mode == "input" and plugin.name.upper() == block):
-								attr[list(plugin.info.attrib.keys())[0]] = atnm
+								keys = list(plugin.info.attrib.keys())
+								if(len(parts) - 1 <= len(keys)):
+									for i in range(0, len(parts)-1):
+										attr[keys[i]] = parts[i+1]
+								else:
+									print("warning: python-droit: plugin '" + block + "' takes a maximum of " + str(len(keys)) + " parameters, " + str(len(parts)-1) + " given.")
 					else:
 						if(block == "INP"):
-							attr["var"] = atnm
+							attr["var"] = parts[1]
 						elif(block == "SIMT"):
-							attr["limit"] = atnm
+							attr["limit"] = parts[1]
 						elif(block == "TEXT"):
-							attr["not"] = atnm
+							attr["not"] = parts[1]
 						elif(block == "SESSION"):
-							attr["cmd"] = atnm
+							attr["cmd"] = parts[1]
 					
 					rule.input.append(_models.DroitRuleInput(block, attr, children))
 				else:
 					rule.input.append(_models.DroitRuleInput(inpx[0], {}, children))
 			
 			for outx in out:
-				outx = outx.split("!")
+				outx = nonEscapeSplit(outx, "!")
 				if(outx[1] != ""):
 					if(outx[0].upper() == "EVAL"):
-						children = [outx[1].replace("&arz;", "!").replace("&dpp;", ":")]
+						children = [escapeCharacters(outx[1])]
 					else:
-						children = outx[1].replace("&arz;", "!").replace("&dpp;", ":").split(",")
+						children = escapeCharacters(outx[1]).split(",")
 				else:
 					children = []
 				rule.output.append(_models.DroitRuleOutput(outx[0], children))
 
 			rules.append(rule)	
+		
+		elif(info != "comment" and warnings and not(legacyValid)):
+			print("warning: python-droit: parseScript: " + info)
 
 	return rules
-
-
-def parseScriptInfoString(string: str) -> dict:
-	data = string.split("\n")
-	ret = {}
-	for line in data:
-		if(len(line) > 3):
-			if(line[0] == "@" and " " in line):
-				parts = line.split(" ")
-				if(len(parts) > 0):
-					if(" ".join(parts[1:]).replace(" ", "") != ""):
-						ret[parts[0][1:]] = " ".join(parts[1:])
-	
-	return ret